pyproject.toml
[project]
name = "pk-common"
version = "0.1.0"
description = "Open-source, neutral API for PicoKey devices"
authors = [
    { name = "PicoKey Community" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyusb>=1.2.1",
    "pyscard>=2.0.7",
    "fido2>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
pk-cli = "pkcommon.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
